DRY find_hosts_by_staleness (#624)
