+++ lib/host_repository.py
@@ -15,10 +15,12 @@ from lib.db import session_guard
 
 __all__ = (
     "add_host",
+    "ALL_STALENESS_STATES",
     "canonical_facts_host_query",
     "create_new_host",
     "find_existing_host",
     "find_host_by_canonical_facts",
+    "find_hosts_by_staleness",
     "stale_timestamp_filter",
     "update_existing_host",
 )
@@ -32,7 +34,7 @@ AddHostResults = Enum("AddHostResults", ["created", "updated"])
 # the priority.
 ELEVATED_CANONICAL_FACT_FIELDS = ("insights_id", "subscription_manager_id")
 
-STALENESS_STATES = ("fresh", "stale", "stale_warning", "unknown")
+ALL_STALENESS_STATES = ("fresh", "stale", "stale_warning", "unknown")
 NULL = None
 
 logger = get_logger(__name__)
@@ -77,19 +79,11 @@ def _find_host_by_elevated_ids(account_number, canonical_facts):
     return None
 
 
-def exclude_culled_hosts(query):
-    config = inventory_config()
-    staleness_conditions = tuple(staleness_to_conditions(config, STALENESS_STATES, stale_timestamp_filter))
-    if "unknown" in STALENESS_STATES:
-        staleness_conditions += (Host.stale_timestamp == NULL,)
-    return query.filter(or_(*staleness_conditions))
-
-
 def canonical_fact_host_query(account_number, canonical_fact, value):
     query = Host.query.filter(
         (Host.account == account_number) & (Host.canonical_facts[canonical_fact].astext == value)
     )
-    return exclude_culled_hosts(query)
+    return find_hosts_by_staleness(ALL_STALENESS_STATES, query)
 
 
 def canonical_facts_host_query(account_number, canonical_facts):
@@ -100,7 +94,7 @@ def canonical_facts_host_query(account_number, canonical_facts):
             | Host.canonical_facts.comparator.contained_by(canonical_facts)
         )
     )
-    return exclude_culled_hosts(query)
+    return find_hosts_by_staleness(ALL_STALENESS_STATES, query)
 
 
 def find_host_by_canonical_fact(account_number, canonical_fact, value):
@@ -129,6 +123,16 @@ def find_host_by_canonical_facts(account_number, canonical_facts):
         logger.debug("Found existing host using canonical_fact match: %s", host)
 
     return host
+
+
+def find_hosts_by_staleness(staleness, query):
+    logger.debug("find_hosts_by_staleness(%s)", staleness)
+    config = inventory_config()
+    staleness_conditions = tuple(staleness_to_conditions(config, staleness, stale_timestamp_filter))
+    if "unknown" in staleness:
+        staleness_conditions += (Host.stale_timestamp == NULL,)
+
+    return query.filter(or_(*staleness_conditions))
 
 
 @metrics.new_host_commit_processing_time.time()
